requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
confusable-homoglyphs==3.2.0
Django==4.0.5
django-crispy==1.8.1
django-crispy-forms==1.14.0
django-registration==3.3
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
